legal_document_analyzer/requirements.txt

pymupdf
python-docx
pillow
pytesseract
spacy
transformers
torch
fastapi
uvicorn
streamlit
pyyaml
pandas
numpy
